refactor: Separate model area state and function into seperate hook.
